Add files via upload
Added some styling to button. Got "Results' button to show but it's buggy, need to fix. . I still need to write if/else for correct/wrong display. Do I need somewhere to store the correct answers?  I am thinking of an array that will store them if the correct button is selected but maybe try to tie correct answer selected, to a div for display. Still need 'Restart' button to function.